Renamed Trait demos to stories to match the rest of the project.
