+++ src/traits/aligned/stories.styles.js
