+++ src/traits/bordered/stories.jsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { jsx } from '@emotion/core'
 
 import * as Story from '_story'
-import * as Styled from './demo.styles'
+import * as Styled from './stories.styles'
 
 const Demo = () => (
   <>
