+++ src/traits/bordered/stories.styles.js
