+++ src/traits/gridded/stories.jsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import { jsx } from '@emotion/core'
 
-import * as Styled from './demo.styles'
+import * as Styled from './stories.styles'
 
 const cols = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]
