+++ src/traits/inline/stories.jsx
@@ -5,7 +5,7 @@ import { jsx } from '@emotion/core'
 
 import * as Story from '_story'
 
-import * as Styled from './demo.styles'
+import * as Styled from './stories.styles'
 
 const Demo = () => {
   return (
