+++ src/traits/traits.stories.jsx
@@ -3,18 +3,18 @@ import React from 'react'
 
 import { A } from '_story'
 
-export { default as Aligned } from './aligned/demo'
-export { default as Bordered } from './bordered/demo'
-export { default as Chromatic } from './chromatic/demo'
-export { default as Contained } from './contained/demo'
-export { default as Flexible } from './flexible/demo'
-export { default as Gridded } from './gridded/demo'
+export { default as Aligned } from './aligned/stories'
+export { default as Bordered } from './bordered/stories'
+export { default as Chromatic } from './chromatic/stories'
+export { default as Contained } from './contained/stories'
+export { default as Flexible } from './flexible/stories'
+export { default as Gridded } from './gridded/stories'
 export { default as Inline } from './inline/stories'
-export { default as Interactive } from './interactive/demo'
-export { default as Margined } from './margined/demo'
-export { default as Responsive } from './responsive/demo'
-export { default as Rounded } from './rounded/demo'
-export { default as Typographic } from './typographic/demo'
+export { default as Interactive } from './interactive/stories'
+export { default as Margined } from './margined/stories'
+export { default as Responsive } from './responsive/stories'
+export { default as Rounded } from './rounded/stories'
+export { default as Typographic } from './typographic/stories'
 
 export const Index = () => (
   <>
